Move ScrollArea and flag enums into ScrollArea.h
